requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
